Renamed Blit to DrawFullscreen feature
